Remove negative lookbehind regexp logic.
As reported in https://github.com/mrodrig/json-2-csv/issues/197, many
browsers don't support negative look behind RegExps, which causes both
calls to this module and dependency modules to fail when that logic was
invoked. In order to fix that, this commit reworks the RegExp to be use
a string character parser to find first index where a non-escaped '.'
character is found.